refactor: move individual functions
